Feature/gnss driver refactor (#28)
* feature:add gnss driver files

not sure that it works

* chore:reallocate in .ld uninitialized data (.bss)

Move them to RAM_D1 to free up DTCMRAM which was the default of CubeMX

* fix:enable RX TX swapping

UART5 global interrupt disabled and it still works.

* feature: IDLE event detection

Description: enhanced the RxEventCallback, added printing function,
proper disable of HT and TC interrupts

* feature: add notification logic inside the RxEventCallback

Description: now gnsstask only runs when there are data to process and
then starts again the DMA...could loose some data if the processing is
very slow...we will see

* feature: check most of the drivers functions

* feature: driver functions are working as expected

Tested Functions: configureNMEATalkerID, configureMessageType,
configureGNSSNavigationMode, configureSystemPositionRate,
setFactoryDefaults, querySoftwareVersion, configureSystemPowerMode

* feature: tried add ACK/NACK functionality on controlGNSS function - not
working

* fix: ack function is working well

checked with multiple commands

* feature: added a small delay in the controlGNSS function to avoid
blocking the mcu

* feature: ACK/NACK with notification - new era

* feature/fix: double buffering logic, finalize control

Description: Implemented a retry loop for GNSS message transmission.
This logic allows the system to make multiple attempts before reporting
an error. Furthermore the first command was a failure for an unknown
reason. Now this is fixed

* feature: receiving messages from GNSS with queue

ACK/NACK not working yet

* feature: function that keeps active only the RMC

Description: The control function works well...needs some extra tuning
though. Also, the ISR needs further investigation

* feature: Configure ISR, Notify Timeouts and switch

Added better control on the ISR, configure timouts for the queues and
the notifications. Also more efficient logic on the control function
when searching for the ACK and NACK

* minor: add log error queue timeout

* chore: add of the NMEA libraries

* feature: add parser function for RMC and GGA

* feature: add init func, enhancement of the parser

* feature: successful integration of GNSS

Description: add a small delay on TMP117 task to avoid wrong readings
from the TMP and the INA

* feature: GNSS output every 5s

* feature: conversion from ddm to decimal degrees

Updates: experimented with precision and concluded that calculations must be
performed using double precision. However, the LLA will be transmitted
as an int32 with 7-digit precision, which is quite accurate considering
that GNSS provides 5-digit precision for the minutes.

* feature: gnss parser almost done except UTC

Description: added filtering based on quality factor, increased the
precision on the ddm to dd conversion

* feature: comment the filtering of the qi for debug