Priliminary transceiver control (#37)
* fix:TX AMP is now initially closed from CubeMX

EN_PA_UHF set GPIO output level to HIGH

* feature: task separation, work in progress

* feature: add some configs of the baseband core

* feature: add freq synthesizer and aux config

* saving progress

* feature: saving progress

Description: tested the frequency synthesizer settings along with the
baseband core set up functions. Also set the rx and tx amp to be closed
by default (from cubemx). With the currect config received frame start
interrupt is detected which I guess causes the AT to get into TXPREP
some times.

* refactor: refactoring all the drivers functions

Description: The core set of driver functions is functioning as expected. Details of the tested functions can be found in the Excel file located on the drive.

* refactor: finalized configs

* test: leakage voltage test on 3V3RF

* feat: notification logic on RF-RX

Description: notification logic on RX with two notifications events for
debugging purposes. Also added a function that returns the received
length with etl expected format and changed the scale of the INA to mV
and mA

* Add notifications - tx not working yet

* saving progress

* feat: TX is working as expected with notifications

BUG:  RF_RST has to be toggled and left to HIGH in order for the transceiver to send tx packets nominally

* feat: sychronization with mutex between rf tasks

* wip: change the tx task delay

* test: increase the tx power to 16 dBm